Modules (#1)
Implement ECMAScript modules support.

The API is not stable at this moment and it is also not documented mostly for this reason.

Apart from that every single part of the current specification as related to ESM has been taken into account and implemented. 

Including top level await, dynamic imports and namespaces.