Improve ambiguous import error and position fixes
This splits the errors to distinguish between ambiguous imports - ones
where multiple of the same identifiers were exported, and we can't know
which it import. And the case where there is just no such export when it
is imported.

The messages were inspired from deno and node.js.

Also fixes line number where it is imported.

The source file is still listed twice.

Fixes #11